+++ vitis/software/Baremetal/test/uz/uz_matrix/test_uz_matrix_add_to_matrix.c
@@ -14,8 +14,8 @@ void tearDown(void)
 }
 
 void test_uz_matrix_add_to_matrix_column_vector(void){
-    size_t rows=1;
-    size_t columns=4;
+    uint32_t rows=1;
+    uint32_t columns=4;
     float A_data[4]={1,2,3,4};
     float C_data[4]={1,2,3,4};
     uz_matrix_t* A=init_array_test_helper(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
@@ -29,8 +29,8 @@ void test_uz_matrix_add_to_matrix_column_vector(void){
 }
 
 void test_uz_matrix_add_to_matrix_row_vector(void){
-    size_t rows=4;
-    size_t columns=1;
+    uint32_t rows=4;
+    uint32_t columns=1;
     float A_data[4]={1,2,3,4};
     float C_data[4]={1,2,3,4};
     uz_matrix_t* A=init_array_test_helper(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
@@ -44,8 +44,8 @@ void test_uz_matrix_add_to_matrix_row_vector(void){
 }
 
 void test_uz_matrix_add_to_matrix_2_times2(void){
-    size_t rows=2;
-    size_t columns=2;
+    uint32_t rows=2;
+    uint32_t columns=2;
     float A_data[4]={1,2,3,4};
     float C_data[4]={1,2,3,4};
     uz_matrix_t* A=init_array_test_helper(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
