+++ vitis/software/Baremetal/test/uz/uz_matrix/test_uz_matrix_elementwise.c
@@ -14,8 +14,8 @@ void tearDown(void)
 }
 
 void test_uz_matrix_vector_elementwise_multiplication_3_times_1(void){
-    size_t rows=3;
-    size_t columns=1;
+    uint32_t rows=3;
+    uint32_t columns=1;
     float A_data[3]={1,2,3};
     float B_data[3]={1,2,3};
     float C_data[3]={134513};
@@ -30,8 +30,8 @@ void test_uz_matrix_vector_elementwise_multiplication_3_times_1(void){
 }
 
 void test_uz_matrix_matrix_elementwise_multiplication_2_times_2(void){
-    size_t rows=2;
-    size_t columns=2;
+    uint32_t rows=2;
+    uint32_t columns=2;
     // | 1 2 | .* | 1 2 | = | 1 4  |
     // | 3 4 |    | 3 4 |   | 9 16 |
     float A_data[4]={1,2,3,4};
@@ -49,8 +49,8 @@ void test_uz_matrix_matrix_elementwise_multiplication_2_times_2(void){
 }
 
 void test_uz_matrix_matrix_elementwise_multiplication_1_times_4(void){
-    size_t rows=1;
-    size_t columns=4;
+    uint32_t rows=1;
+    uint32_t columns=4;
     float A_data[4]={1,2,3,4};
     float B_data[4]={1,2,3,4};
     float C_data[4]={21345};
