+++ vitis/software/Baremetal/test/uz/uz_matrix/uz_matrix_init_helper.c
@@ -2,7 +2,7 @@
 
 #include <stdlib.h>
 
-uz_matrix_t *init_array_test_helper(float* data, size_t length_of_data,size_t rows,size_t columns)
+uz_matrix_t *init_array_test_helper(float* data, uint32_t length_of_data,uint32_t rows,uint32_t columns)
 {
     uz_matrix_t *pointer_to_matrix = malloc(sizeof(uz_matrix_t));
     uz_matrix_init(pointer_to_matrix,data,length_of_data,rows,columns);
